fix(list view): Headers must be on own row
Refactored list view to place column headers on their own row. An issue was discovered after placing an info icon in the header. When the user hovers over the icon, a tooltip is displayed, but the first row is unexpectedly highlighted. Moving headers to their own row ensures the mouseover event is not fired for the first list view row. This update will ultimately be used with the list view on the codebases page.